@@ -2,17 +2,17 @@
   "name": "api",
   "version": "0.0.1",
   "description": "",
-  "author": "zan.remskar8@gmail.com",
+  "author": "",
   "license": "MIT",
   "scripts": {
     "dev": "nodemon",
-    "build": "tsc -p tsconfig.build.json",
-    "format": "prettier --write \"src/**/*.ts\"",
+    "prebuild": "rimraf dist",
+    "build": "nest build",
+    "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "start": "ts-node -r tsconfig-paths/register src/main.ts",
-    "start:dev": "nodemon",
-    "start:debug": "nodemon --config nodemon-debug.json",
-    "prestart:prod": "rimraf dist && npm run build",
-    "start:prod": "node dist/main.js",
+    "start:dev": "nest start --watch",
+    "start:debug": "nest start --debug --watch",
+    "start:prod": "node dist/main",
     "lint": "tslint -p tsconfig.json -c tslint.json",
     "test": "jest",
     "test:watch": "jest --watch",
@@ -21,39 +21,43 @@
     "test:e2e": "jest --config ./test/jest-e2e.json"
   },
   "dependencies": {
-    "@nestjs/common": "^6.0.0",
-    "@nestjs/config": "^0.2.4",
-    "@nestjs/core": "^6.0.0",
+    "@nestjs/common": "^6.7.2",
+    "@nestjs/core": "^6.7.2",
     "@nestjs/jwt": "^6.1.1",
-    "@nestjs/mongoose": "^6.3.1",
-    "@nestjs/passport": "^6.1.1",
-    "@nestjs/platform-express": "^6.0.0",
-    "@typegoose/typegoose": "^6.2.2",
-    "bcryptjs": "^2.4.3",
-    "mongo-express": "^0.54.0",
-    "mongoose": "^5.9.2",
-    "nestjs-typegoose": "^7.1.0",
-    "passport": "^0.4.1",
+    "@nestjs/mongoose": "^6.1.2",
+    "@nestjs/passport": "^6.1.0",
+    "@nestjs/platform-express": "^6.7.2",
+    "bcrypt": "^3.0.6",
+    "dotenv": "^8.2.0",
+    "gravatar": "^1.8.0",
+    "mongoose": "^5.7.8",
+    "neconfig": "^1.1.0",
+    "nodemon": "^2.0.2",
+    "passport": "^0.4.0",
     "passport-jwt": "^4.0.0",
-    "reflect-metadata": "^0.1.12",
-    "rimraf": "^2.6.2",
-    "rxjs": "^6.3.3"
+    "reflect-metadata": "^0.1.13",
+    "rimraf": "^3.0.0",
+    "rxjs": "^6.5.3"
   },
   "devDependencies": {
-    "@nestjs/testing": "^6.0.0",
-    "@types/express": "^4.16.0",
-    "@types/jest": "^23.3.13",
-    "@types/node": "^10.12.18",
-    "@types/supertest": "^2.0.7",
-    "jest": "^23.6.0",
-    "nodemon": "^1.18.9",
-    "prettier": "^1.15.3",
-    "supertest": "^3.4.1",
-    "ts-jest": "^23.10.5",
-    "ts-node": "^7.0.1",
-    "tsconfig-paths": "^3.7.0",
-    "tslint": "5.12.1",
-    "typescript": "^3.2.4"
+    "@nestjs/cli": "^6.9.0",
+    "@nestjs/schematics": "^6.7.0",
+    "@nestjs/testing": "^6.7.1",
+    "@types/express": "^4.17.1",
+    "@types/jest": "^24.0.18",
+    "@types/mongoose": "^5.5.29",
+    "@types/node": "^12.7.5",
+    "@types/passport-jwt": "^3.0.2",
+    "@types/supertest": "^2.0.8",
+    "jest": "^24.9.0",
+    "prettier": "^1.18.2",
+    "supertest": "^4.0.2",
+    "ts-jest": "^24.1.0",
+    "ts-loader": "^6.1.1",
+    "ts-node": "^8.4.1",
+    "tsconfig-paths": "^3.9.0",
+    "tslint": "^5.20.0",
+    "typescript": "^3.6.3"
   },
   "jest": {
     "moduleFileExtensions": [
@@ -66,7 +70,7 @@
     "transform": {
       "^.+\\.(t|j)s$": "ts-jest"
     },
-    "coverageDirectory": "../coverage",
+    "coverageDirectory": "./coverage",
     "testEnvironment": "node"
   }
 }
