@@ -9,26 +9,74 @@ import { StatusBar } from '@ionic-native/status-bar/ngx';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 
-import { SocketIoModule, SocketIoConfig } from 'ngx-socket-io';
-import { HttpClientModule } from '@angular/common/http';
+import { IonicStorageModule, Storage } from '@ionic/storage';
+import { HttpClientModule, HttpClient  } from '@angular/common/http';
 
 import { SecureStorage} from '@ionic-native/secure-storage/ngx';
-import { NativeStorage } from '@ionic-native/native-storage/ngx';
 
-const config: SocketIoConfig = { url: 'http://localhost:3000', options: {}};
+import { TOKEN_KEY } from './services/api.service';
+import { JwtModule, JWT_OPTIONS } from '@auth0/angular-jwt';
 
+export function jwtOptionsFactory(storage) {
+  return {
+    tokenGetter: () => {
+      return storage.get(TOKEN_KEY);
+    },
+    whitelistedDomains: ['localhost:5000'] // Add your Heroku URL in here!
+  }
+}
+
+//TRANSLATION
+
+import { TranslateModule, TranslateLoader, TranslateCompiler } from '@ngx-translate/core';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import {TranslateMessageFormatCompiler} from 'ngx-translate-messageformat-compiler';
+
+export function LanguageLoader(http: HttpClient) {
+  return new TranslateHttpLoader(http, 'assets/i18n/', '.json');
+}
+
+export function HttpLoaderFactory(http: HttpClient) {
+  return new TranslateHttpLoader(http);
+}
 
 @NgModule({
   declarations: [AppComponent],
   entryComponents: [],
-  imports: [BrowserModule, IonicModule.forRoot(), AppRoutingModule, SocketIoModule.forRoot(config), HttpClientModule],
+  imports: [
+    IonicModule.forRoot(),
+    IonicStorageModule.forRoot(),
+
+    JwtModule.forRoot({
+      jwtOptionsProvider: {
+        provide: JWT_OPTIONS,
+        useFactory: jwtOptionsFactory,
+        deps: [Storage]
+      }
+    }),
+    
+    TranslateModule.forRoot({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: HttpLoaderFactory,
+        deps: [HttpClient]
+      },
+      compiler: {
+        provide: TranslateCompiler,
+        useClass: TranslateMessageFormatCompiler
+      }
+    }),
+
+    BrowserModule,
+    AppRoutingModule,
+    HttpClientModule,
+    HttpClientModule
+    ],
   providers: [
     StatusBar,
     SplashScreen,
     SecureStorage,
-    NativeStorage,
-    { provide: RouteReuseStrategy, useClass: IonicRouteStrategy },
-    
+    { provide: RouteReuseStrategy, useClass: IonicRouteStrategy }
   ],
   bootstrap: [AppComponent]
 })
