@@ -5,15 +5,15 @@ import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 
 import { HomePage } from './home.page';
-import {ComponentsModule} from '../../components/components.module';
 import { SecureStorage} from '@ionic-native/secure-storage/ngx';
+import { TranslateModule } from '@ngx-translate/core';
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
     IonicModule,
-    ComponentsModule,
+    TranslateModule.forChild(),
     RouterModule.forChild([
       {
         path: '',
