@@ -8,6 +8,12 @@ import { Router } from "@angular/router";
 
 import { SecureStorage, SecureStorageObject } from '@ionic-native/secure-storage/ngx';
 
+import { ApiService, User } from '../../services/api.service';
+
+import { TranslateConfigService } from '../../services/translate-config.service';
+
+import {TranslateService} from '@ngx-translate/core';
+ 
 
 @Component({
   selector: 'app-home',
@@ -19,49 +25,105 @@ export class HomePage {
   dogName:string = "/";
   dogBreed:string = "/";
   todos:Array<string> = [];
+  user:User;
 
-  constructor(private alert:AlertController, private httpReq: HttpRequestService, private modalController: ModalController, private router:Router, private secureStorage: SecureStorage){
-    this.todos = this.httpReq.getTodos();
-  }
+  selectedLanguage:string;
+  todo_label:string = "";
+
+  constructor(
+    private alert:AlertController,
+    private api:ApiService, 
+    private httpReq: HttpRequestService, 
+    private modalController: ModalController,
+    private router:Router, 
+    private secureStorage: SecureStorage,
+    private translateConfigService: TranslateConfigService,
+    private translate:TranslateService
+    ){}
 
   ngOnInit(){
     console.log("hello from home page");
 
-    this.dogName = this.httpReq.getDogData().dogName;
-    this.dogBreed = this.httpReq.getDogData().dogBreed;
-    
+    this.loadUserData();
+    // this.selectedLanguage = this.translateConfigService.getDefaultLanguage();
+    this.translateConfigService.setLanguage("sl");
+
+ 
   }
 
+  loadUserData(){
+    this.api.getUserData().subscribe(a =>{
+      this.user = a;
+
+      if(a.dogBreed == "" && a.dogName == "" || a.dogBreed == null && a.dogName == null){
+        this.openAddDog();
+      }else{
+        this.dogBreed = this.user.dogBreed;
+        this.dogName = this.user.dogName;
+        this.todos = this.user.todos;
+      }
+
+    })
+  }
+  
   async openTOdo(){
 
-      const alert = await this.alert.create({
-        header: `Please add task`,
-        inputs:[
-          {
-            name: "todo",
-            type: "text"
-          }
-        ],
-          buttons: [
-            {
-              text: 'Ok',
-              handler: (a) => {
-                this.todos.push(a.todo);
-                this.httpReq.saveUserTodo(this.todos);
+    const alert = await this.alert.create({
+      header: this.translate.instant('HOME.todo_label'),
+      inputs:[
+        {
+          name: "todo",
+          type: "text"
+        }
+      ],
+      buttons: [
+        {
+          text: 'Ok',
+          handler: (a) => {
+            this.todos.push(a.todo);
+                this.api.updateUser(this.user._id, {"todos": this.todos}).subscribe(a => {
+                  console.log(a);
+                })
               }
             }
           ]})
           await alert.present();
-  }
-
+        }
+        
   logOut(){
-    this.router.navigate(['login']);
+    this.router.navigate(['']);
   }
 
-  delete(i:number){
-    this.todos.splice(i, 1);
+  signOut() {
+    this.api.logout();
   }
-
   
+  delete(todo){
+    let index = this.todos.indexOf(todo);
 
+    if(index > -1){
+      this.todos.splice(index, 1);
+      this.api.updateUser(this.user._id, {"todos": this.todos}).subscribe(a => {
+        console.log(a);
+      })
+    }
+
+  }
+
+  async openAddDog(){
+    const alert = await this.alert.create({
+      header: `Would you like to add a dog?`,
+        buttons: [
+          {
+            text: 'Yes',
+            handler: () => {
+              this.router.navigateByUrl('/dog-form');
+            }
+          },
+          {text: "No"}
+        ]})
+        await alert.present();
+  }
+  
+  
 }
