@@ -1,3 +1,6 @@
 export const environment = {
-  production: true
+  production: true,
+  apiUrl: 'api' 
+  //api/
+  // Use your Heroku or hosting URL instead
 };
